feat: adjust admin and instrutor register

+++ api/src/main/java/sync/fit/api/dto/response/RelatorioSalarioInstrutorDTO.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class RelatorioSalarioInstrutorDTO {
     private String nome;
-    private Double salario;
+    private BigDecimal salario;
 }
